requirements.txt
google-genai>=0.2.0
google-generativeai>=0.8.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0